Move editor messages out of tooling-models
