+++ subprojects/provider/src/main/kotlin/org/gradle/kotlin/dsl/resolver/EditorMessages.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.kotlin.dsl.resolver
+
+
+object EditorMessages {
+
+    const val failure = "Script dependencies resolution failed"
+    const val failureUsingPrevious = "Script dependencies resolution failed, using previous dependencies"
+
+    private
+    const val forMoreInformation = "run 'gradle tasks' for more information"
+
+    const val buildConfigurationFailed = "Build configuration failed, $forMoreInformation"
+    const val buildConfigurationFailedInCurrentScript = "This script caused build configuration to fail, $forMoreInformation"
+
+
+    fun defaultErrorMessageFor(cause: Throwable) =
+        "${cause::class.java.name}, $forMoreInformation"
+}
