+++ subprojects/tooling-builders/src/main/kotlin/org/gradle/kotlin/dsl/tooling/builders/EditorReportsBuilder.kt
@@ -16,10 +16,11 @@
 
 package org.gradle.kotlin.dsl.tooling.builders
 
-import org.gradle.kotlin.dsl.tooling.models.EditorMessages
 import org.gradle.kotlin.dsl.tooling.models.EditorPosition
 import org.gradle.kotlin.dsl.tooling.models.EditorReport
 import org.gradle.kotlin.dsl.tooling.models.EditorReportSeverity
+
+import org.gradle.kotlin.dsl.resolver.EditorMessages
 
 import org.gradle.internal.exceptions.LocationAwareException
